An Itty Bitty GUI Addition
Energy is now able to be represented by an Energy Tank. While crudely
added at the moment, it could be possible to have a system where greater
amounts of 'Energy Tanks' can be added through various means. For the
player, they're only useful for Magical Actions, like casting spells or
defending one's self.

Currently, the GUI system is very, veeeery basic. The little bar will
move with the change in energy and it will deplete, much like the energy
tanks in StreetPass Mansion, the inspiration for the combat here. The
issue is that the system is only going to use one tank. I'd prefer to
represent the energy with multiple reserves to give something of a
visual cue to the player that their equipment has granted or taken an
amount of energy from them. This let's them gauge how reckless they can
be in an instant.

I dunno, a standard bar seems very...flat. A bar can grow an shrink but
to a new player, that seems a little arbitrary.

A visual 'Tank' on the other hand is much more likely to convey an idea
of the units they have to work with because now, they have a context.

Instead of "OK, that took off some of my bar. I think I have enough left
for another move", it's "OK, that emptied an entire Energy Tank. I know
that my next move takes up two tanks and I only have one full one right
now. I need to turtle up and recharge."

I hope that makes sense. Tori, Foxy, poke me what you think.