A Snag
In an effort to expand on the GUI Energy element, I'd hit a confusing
road block.

The Array List that is holding each Energy Tank is reporting Null
Pointer Exceptions even though I am quite certain that I have
initialised both the list and it's contents properly. So, the reason why
it's doing this is irritating me greatly.

Hopefully, this can get figured out soon.